Make sure to call max on identically-typed arguments.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@16492 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objmgr/split/blob_splitter_impl.cpp
@@ -187,7 +187,7 @@ void CBlobSplitterImpl::CollectPieces(const CSeq_annot_SplitInfo& info)
 void CBlobSplitterImpl::Add(const SAnnotPiece& piece)
 {
     EAnnotPriority priority = piece.m_Priority;
-    m_Pieces.resize(max(m_Pieces.size(), priority+1u));
+    m_Pieces.resize(max(m_Pieces.size(), priority + size_t(1)));
     if ( !m_Pieces[priority] ) {
         m_Pieces[priority] = new CAnnotPieces;
     }
@@ -380,6 +380,9 @@ END_NCBI_SCOPE
 /*
 * ---------------------------------------------------------------------------
 * $Log$
+* Revision 1.10  2004/06/30 23:27:59  ucko
+* Make sure to call max on identically-typed arguments.
+*
 * Revision 1.9  2004/06/30 20:56:32  vasilche
 * Added splitting of Seqdesr objects (disabled yet).
 *
